Ad image 2 text, fix broken semantic chunk
